refactor (dual-servicing) : api as router and make views as static folder path
