--- src/router/view.js
@@ -1,3 +1,0 @@
-const router = require('express').Router();
-
-module.exports = router;
